Fixed gamemode name collision
Fixed the bug where the (wrapper) gamemode code was loaded by the
bootloader. This was because the recursive function explored the folder
jailbreak in the virtual lua folder and recursed into the gamemode.
Gamemode has been renamed to jailbreak_potatoman. Probably should pick a
different/better name in the future or include all files without
recursion